# docs/swagger.yaml
definitions:
  controller.response:
    properties:
      message:
        type: string
    type: object
...
  models.Parameters:
    properties:
      created:
        $ref: '#/definitions/models.Admins'
      createdAt:
        type: string
      id:
        type: integer
      nameEn:
        type: string
      nameRu:
        type: string
      nameUz:
        type: string
      position:
        type: integer
    type: object
...
  models.Products:
    properties:
      brand:
        $ref: '#/definitions/models.Brand'
      brandId:
        type: integer
      country:
        $ref: '#/definitions/models.Country'
      countryId:
        type: integer
      created:
        $ref: '#/definitions/models.Admins'
      createdAt:
        type: string
      deleted:
        $ref: '#/definitions/models.Admins'
      deletedAt:
        type: string
      descriptionEn:
        type: string
      descriptionRu:
        type: string
      descriptionUz:
        type: string
      id:
        type: integer
      image:
        type: string
      isActive:
        type: boolean
      isNew:
        type: boolean
      isTop:
        type: boolean
      nameEn:
        type: string
      nameRu:
        type: string
      nameUz:
        type: string
      parent:
        $ref: '#/definitions/models.Category'
      parentId:
        type: integer
      position:
        type: integer
      price:
        type: number
      seoDescription:
        type: string
      seoTitle:
        type: string
      updated:
        $ref: '#/definitions/models.Admins'
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server CRM server.
  title: Golang CRM Swagger Documentation
  version: "1.0"
paths:
...
  /api/analog/products/{id}:
    delete:
      consumes:
      - application/json
      description: this api is  Add analog products
      parameters:
      - description: news id
        in: path
        name: id
        required: true
        type: integer
      - description: new products add
        in: query
        name: products
        required: true
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Add analog products
      tags:
      - Analogs
    get:
      consumes:
      - application/json
      description: this api is  Add analog products
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Products'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Add analog products
      tags:
      - Analogs
    put:
      consumes:
      - application/json
      description: this api is  Add analog products
      parameters:
      - description: news id
        in: path
        name: id
        required: true
        type: integer
      - description: new products add
        in: query
        name: products
        required: true
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Add analog products
      tags:
      - Analogs
...
  /api/parameter/{id}:
    delete:
      consumes:
      - application/json
      description: this api is to delete parameter
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Delete parameter
      tags:
      - Parameter
    get:
      consumes:
      - application/json
      description: this api is to get Parameter
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Parameters'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Get Parameter
      tags:
      - Parameter
    put:
      consumes:
      - application/json
      description: this api is Update parameter
      parameters:
      - description: update id
        in: path
        name: id
        required: true
        type: string
      - in: query
        name: nameEn
        type: string
      - in: query
        name: nameRu
        type: string
      - in: query
        name: nameUz
        type: string
      - in: query
        name: position
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Parameters'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Update parameter
      tags:
      - Parameter
  /api/product:
    get:
      consumes:
      - application/json
      description: this api is get product
      parameters:
      - in: query
        name: isNew
        type: boolean
      - in: query
        name: isTop
        type: boolean
      - in: query
        name: multiSearch
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: parentId
        type: integer
      - in: query
        name: priceFrom
        type: number
      - in: query
        name: priceTo
        type: number
      - description: product brand ids
        in: query
        name: brandId
        type: array
      - description: product country ids
        in: query
        name: countryId
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      summary: get product
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: this api is create new product
      parameters:
      - in: formData
        name: brandId
        type: integer
      - in: formData
        name: countryId
        type: integer
      - in: formData
        name: descriptionEn
        type: string
      - in: formData
        name: descriptionRu
        type: string
      - in: formData
        name: descriptionUz
        type: string
      - in: formData
        name: isActive
        type: boolean
      - in: formData
        name: isNew
        type: boolean
      - in: formData
        name: isTop
        type: boolean
      - in: formData
        name: nameEn
        type: string
      - in: formData
        name: nameRu
        type: string
      - in: formData
        name: nameUz
        type: string
      - in: formData
        name: parentId
        type: integer
      - in: formData
        name: position
        type: integer
      - in: formData
        name: price
        type: number
      - in: formData
        name: seoDescription
        type: string
      - in: formData
        name: seoTitle
        type: string
      - description: file
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Products'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Create new product
      tags:
      - Product
...
  /api/product/all:
    get:
      consumes:
      - application/json
      description: this api is get product
      parameters:
      - in: query
        name: isNew
        type: boolean
      - in: query
        name: isTop
        type: boolean
      - in: query
        name: multiSearch
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: parentId
        type: integer
      - in: query
        name: priceFrom
        type: number
      - in: query
        name: priceTo
        type: number
      - description: product brand ids
        in: query
        name: brandId
        type: array
      - description: product country ids
        in: query
        name: countryId
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: get product
      tags:
      - Product
  /api/product/list:
    post:
      consumes:
      - application/json
      description: this api is get product
      parameters:
      - description: product brand ids
        in: body
        name: data
        schema:
          $ref: '#/definitions/models.ProductsIds'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProductsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      summary: get product
      tags:
      - Product
...
  /api/product/parameter/{id}:
    post:
      consumes:
      - application/json
      description: this api will create gotten parameters and delete other relations
        product parameters
      parameters:
      - description: product id
        in: path
        name: id
        required: true
        type: integer
      - description: data body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.ProductParamReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Products'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Create product parameters
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    description: AUTH.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"